# Allocate map for results to hold expected number of keys